+++ tests/resources/ethereumetl_airflow/test_parse/dydx/SoloMargin_event_LogTrade.json_True_1.sql
