+++ tests/resources/ethereumetl_airflow/test_parse/ens/Registrar0_event_NewBid.json_False_1.sql
@@ -8,14 +8,14 @@ WITH parsed_logs AS
     ,`blockchain-etl.ethereum_ens_internal.parse_Registrar0_event_NewBid`(logs.data, logs.topics) AS parsed
 FROM `bigquery-public-data.crypto_ethereum.logs` AS logs
 WHERE address in (
-
+    
     '0x6090a6e47849629b7245dfa1ca21d94cd15878ef'
-
+    
   )
   AND topics[SAFE_OFFSET(0)] = '0xb556ff269c1b6714f432c36431e2041d28436a73b6c3f19c021827bbdc6bfc29'
-
+  
   AND DATE(block_timestamp) = '2020-01-01'
-
+  
   )
 SELECT
      block_timestamp
@@ -23,7 +23,7 @@ SELECT
      ,transaction_hash
      ,log_index
      ,contract_address
-
+     
     ,parsed.hash AS `hash`
     ,parsed.bidder AS `bidder`
     ,parsed.deposit AS `deposit`
