+++ tests/resources/ethereumetl_airflow/test_parse/ens/Registrar0_event_NewBid.json_False_3.sql
