+++ tests/resources/ethereumetl_airflow/test_parse/idex/Exchange_call_trade.json_True_1.sql
