+++ tests/resources/ethereumetl_airflow/test_parse/uniswap/Uniswap_event_AddLiquidity.json_True_1.sql
